Unified projection tests
